construction_management/requirements.txt
django==5.0
djangorestframework==3.15
mysqlclient==2.2.4
python-decouple==3.8
django-cors-headers==4.4.0
